+++ inc/includes/localization.php
@@ -2,8 +2,8 @@
 /**
  * @Author: Timi Wahalahti
  * @Date:   2019-12-03 11:03:31
- * @Last Modified by:   Timi Wahalahti
- * @Last Modified time: 2021-05-11 14:40:19
+ * @Last Modified by:   Roni Laukkarinen
+ * @Last Modified time: 2021-06-09 16:11:03
  *
  * @package air-light
  */
@@ -46,6 +46,7 @@ function get_default_localization_strings( $language = 'en' ) {
       'Search'                                       => __( 'Search', 'air-light' ),
       'Block missing required data'                  => __( 'Block missing required data', 'air-light' ),
       'This error is shown only for logged in users' => __( 'This error is shown only for logged in users', 'air-light' ),
+      'No results found for your search'                   => __( 'No results found for your search', 'air-light' ),
     ],
     'fi'  => [
       'Add a menu'                                   => 'Luo uusi valikko',
@@ -65,6 +66,7 @@ function get_default_localization_strings( $language = 'en' ) {
       'Search'                                       => 'Haku',
       'Block missing required data'                  => 'Lohkon pakollisia tietoja puuttuu',
       'This error is shown only for logged in users' => 'Tämä virhe näytetään vain kirjautuneille käyttäjille',
+      'No results for your search'                   => 'Haullasi ei löytynyt tuloksia',
     ],
   ];